pyproject.toml
[project]
name = "agent-for-practical-use"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.97.1",
    "python-dotenv>=1.1.1",
    "wandb>=0.21.0",
    "weave>=0.51.56",
]
